+++ src/main/js/index.tsx
@@ -40,6 +40,8 @@ import RepoEngineConfig from "./workflow/RepoEngineConfig";
 import GlobalEngineConfig from "./workflow/GlobalEngineConfig";
 import ApprovedByXReviewersRuleConfiguration from "./workflow/ApprovedByXReviewersRuleConfiguration";
 import { Branch, Repository } from "@scm-manager/ui-types";
+import PullRequestHitRenderer from "./search/PullRequestHitRenderer";
+import CommentHitRenderer from "./search/CommentHitRenderer";
 
 type PredicateProps = {
   repository?: Repository;
@@ -136,3 +138,5 @@ binder.bind("landingpage.mytask", PullRequestTodos);
 binder.bind("landingpage.mytask", PullRequestReview);
 
 binder.bind("reviewPlugin.workflow.config.ApprovedByXReviewersRule", ApprovedByXReviewersRuleConfiguration);
+binder.bind("search.hit.pullRequest.renderer", PullRequestHitRenderer);
+binder.bind("search.hit.indexedComment.renderer", CommentHitRenderer);